src/llm-server/requirements.txt
accelerate
bitsandbytes
fastapi[all]
google-cloud-storage
InstructorEmbedding
langchain
langchain-community
peft
transformers==4.30
uvicorn
datasets>=2.2.0
pyarrow==8.0.0
jsonlines
numpy
pydantic
requests>=2.26.0
scikit_learn>=1.0.2
scipy
sentence_transformers>=2.2.0
torch
tqdm
rich
